Add better compat aggregation logic.
Still not perfect, but assumes that if there are multiple reports for a
specific configuration and version, the best one should be considered.

This improves skew from:
 * Multiple identical reports from the same exact device.
 * Changed ratings, e.g. Playable -> Perfect from a device.

Also, when multiple compatibility ratings are tied for most common (e.g.
Perfect and Playable), take the better one consistently.  Let's assume
optimistically.